Example of two-axis extension: picojson + variant
The library supports extension of data types (e.g. variant) and also
extension of data formats (e.g. picojson). This leaves a "hole" in the
2d matrix of data types + data formats. This example shows how to fill
in the hole with an extension for a specific data type and a specific
data format. It encodes a variant as a json object {variant: data:}.
The reader is not bulletproof. It checks that the input json has the
necessary fields but does not check that they are the only fields.