Fix #4 - BREAKING - change encoding of char; remove friend access
The C++ standard says char can be signed or unsigned (implementation
defined). Since traverse is used for network messages, and the signed
and unsigned encodings are different, this means a client could encode
where char is signed and a server could decode where a char is
unsigned, and get the wrong result.

To fix this I encoded char, signed char, and unsigned char the same
way as unsigned int. This changes the encoding of char, and thus
breaks any code that relied on it, but shouldn't break the intended
use of Traverse for network messages, as long as client and server are
both updated.

Another breaking change: the friend access macro was not working
reliably across compilers, and I was unable to figure out how to make
it work, so I removed that feature.