Fix #2, undefined behavior
<https://stackoverflow.com/a/37307948> — in theory, -INT_MIN is
sometimes undefined and sometimes defined, but in practice it's
undefined.

When value is int64_t, -value is undefined so -value-1 is undefined.
But -(value+1) is defined and non-negative because value < 0,
and *then* I converted that to uint64_t.

Similarly, when value is uint64_t, -value may be undefined.
<https://stackoverflow.com/q/8026694> says it's defined in C++ but not
C. To avoid issues and also make the code easier to read I converted
it to int64_t.